Updating oxi build & enabling alpha optimisations
