+++ codecs/oxipng/pkg/oxipng.d.ts
@@ -1,8 +1,8 @@
 /* tslint:disable */
 /* eslint-disable */
 /**
-* @param {Uint8Array} data 
-* @param {number} level 
-* @returns {Uint8Array} 
+* @param {Uint8Array} data
+* @param {number} level
+* @returns {Uint8Array}
 */
 export function optimise(data: Uint8Array, level: number): Uint8Array;
